twitter-scraper/requirements.txt
twscrape>=0.17
pandas>=2.0
python-dotenv>=1.0
playwright>=1.40
